+++ magentoerpconnect/CHANGES.rst
@@ -8,6 +8,9 @@ Changelog
 * allow to customize the available versions without overriding the 'version' field
 * Products deactivated on Magento are imported as not sellable / purchasable
 * New option 'Create Invoice On' on payment methods with options 'on paid', 'on validate'
+* Using Magento on PHP 5.4 without using the compatibility patch would
+  break syncs'. Correct solution is to install the patch on Magento
+  though! http://magento.com/blog/magento-news/magento-now-supports-php-54
 
 
 2.3.1 (2014-01-23)